Harshitha/requirements.txt
streamlit
numpy
pandas
PyPDF2
docx2txt
scikit-learn
textblob
nltk
Pillow
